backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.2
eth-account==0.13.4
structlog==24.4.0
tenacity==9.0.0
